src/requirements/site.txt
asgiref==3.7.2
Django==4.2.20
django-unfold==0.55.2
sqlparse==0.4.4
tzdata==2023.3

gunicorn==20.1.0
whitenoise==6.7.0

python-dotenv==1.1.1
python-decouple==3.8

pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1
xlsxwriter==3.2.9
